requirements.txt
Django==4.2.*
psycopg2-binary==2.9.*
django-environ==0.11.*
whitenoise==6.*
django-htmx==1.*

# Producción (evita instalar en Windows):
gunicorn==21.* ; platform_system != "Windows"

# Herramientas de testing y cobertura:
pytest==8.*
pytest-django==4.*
pytest-cov==5.*
coverage==7.*
